split util/slice
